Version 2.0.2: sparse matrices support and minor fix/restyling
Main solvers (lin_uvlm2d_*.py) restyled and support for sparse matrices
added in linear dynamic solution. The state-space matrices are here
converted to sparse after allocation as full. This does not really
affect the code performance, but the time-stepping and the E matrix
inversion both benefit from the use of sparse matrices.

Warning: the allocation of the linear system matrices is inefficient as
based on for loops over non-contingent array elements (e.g. influence
coefficient matrix allocation). Sparsity can not be exploited here as
these matrices are full. For improving performance, for loops should be
rewritten in C++/Fortran and the the allocation modified so as to
manipulate contingent elements.

Other notes/improvements:
- small bug fixed in save.py module.
- motion superimposed in set_dyn.py methods.
- few fixes in test* modules